+++ Runtime/Core/BuiltinMaterials.cs
@@ -54,8 +54,6 @@ namespace UnityEngine.ProBuilder
         /// <seealso cref="pointShader"/>
         public const string dotShader = "Hidden/ProBuilder/VertexShader";
 
-        internal static readonly Color previewColor = new Color(.5f, .9f, 1f, .56f);
-
         static Shader s_SelectionPickerShader;
 
         static bool s_GeometryShadersSupported;
@@ -66,8 +64,6 @@ namespace UnityEngine.ProBuilder
         static Material s_EdgePickerMaterial;
         static Material s_UnityDefaultDiffuse;
         static Material s_UnlitVertexColorMaterial;
-        static Material s_ShapePreviewMaterial;
-
 
 #if HDRP_7_1_0_OR_NEWER
         static string k_EdgePickerMaterial = "Materials/EdgePickerHDRP";
@@ -112,7 +108,7 @@ namespace UnityEngine.ProBuilder
             {
                 Log.Error("VertexPicker material not loaded... please re-install ProBuilder to fix this error.");
                 s_VertexPickerMaterial = new Material(Shader.Find(k_VertexPickerShader));
-        }
+            }
 
             if ((s_EdgePickerMaterial = Resources.Load<Material>(k_EdgePickerMaterial)) == null)
             {
@@ -121,15 +117,6 @@ namespace UnityEngine.ProBuilder
             }
 
             s_UnlitVertexColorMaterial = (Material)Resources.Load("Materials/UnlitVertexColor", typeof(Material));
-
-            s_ShapePreviewMaterial = new Material(s_DefaultMaterial.shader);
-            s_ShapePreviewMaterial.hideFlags = HideFlags.HideAndDontSave;
-
-            if (s_ShapePreviewMaterial.HasProperty("_MainTex"))
-                s_ShapePreviewMaterial.mainTexture = (Texture2D)Resources.Load("Textures/GridBox_Default");
-
-            if (s_ShapePreviewMaterial.HasProperty("_Color"))
-                s_ShapePreviewMaterial.SetColor("_Color", previewColor);
         }
 
         /// <summary>
@@ -303,15 +290,6 @@ namespace UnityEngine.ProBuilder
             {
                 Init();
                 return s_UnlitVertexColorMaterial;
-            }
-        }
-
-        internal static Material ShapePreviewMaterial
-        {
-            get
-            {
-                Init();
-                return s_ShapePreviewMaterial;
             }
         }
     }