+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed rect selection not working with Universal Render Pipeline.
 - [case: 1281254] Fixed shader warning in URP for UNITY_PASS_FORWARDBASE macro redefinition.
 - [case: 1294866] Fix annotation warning on InitializeOnLoad call.
+- [case: 1304561] Fix `Shape Preview` not rendering with the blue preview tint.
 
 ## [4.4.1] - 2020-11-10